pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sigil-experiments"
version = "0.1.0"
description = "Sigil ghost signal training survival experiments"
requires-python = ">=3.12"
dependencies = [
    "sigil-watermark>=0.2.0",
    "torch>=2.0",
    "diffusers>=0.28.0",
    "transformers>=4.38.0",
    "accelerate>=0.28.0",
    "wandb>=0.16.0",
    "safetensors>=0.4.0",
    "datasets>=2.18.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "xformers>=0.0.25",
    # CLI
    "typer>=0.15.0",
    "rich>=13.9.0",
    "runpod>=1.7.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]

[project.scripts]
pod = "runpod_cli.cli:app"

[project.optional-dependencies]
captioning = ["salesforce-lavis>=1.0.0"]
dev = ["pytest", "ruff", "ipython", "jupyter"]

[tool.hatch.build.targets.wheel]
packages = ["src/runpod_cli"]
